+++ src/main/java/com/yoyomo/domain/application/application/dto/request/DataRequest.java
@@ -1,6 +1,6 @@
 package com.yoyomo.domain.application.application.dto.request;
 
-import com.yoyomo.domain.application.domain.model.Reply;
+import com.yoyomo.domain.application.domain.vo.Reply;
 
 public record DataRequest(
         String v
