+++ src/main/java/com/yoyomo/domain/application/application/usecase/ApplicationImportUseCase.java
@@ -2,9 +2,9 @@ package com.yoyomo.domain.application.application.usecase;
 
 import com.yoyomo.domain.application.application.dto.request.ApplicationImportRequest;
 import com.yoyomo.domain.application.domain.entity.Application;
-import com.yoyomo.domain.application.domain.model.ApplicationReply;
 import com.yoyomo.domain.application.domain.service.AnswerSaveService;
 import com.yoyomo.domain.application.domain.service.ApplicationSaveService;
+import com.yoyomo.domain.application.domain.vo.ApplicationReply;
 import com.yoyomo.domain.club.domain.service.ClubManagerAuthService;
 import com.yoyomo.domain.user.domain.entity.User;
 import lombok.RequiredArgsConstructor;
