+++ src/main/java/com/yoyomo/domain/application/domain/service/AnswerSaveService.java
@@ -3,8 +3,8 @@ package com.yoyomo.domain.application.domain.service;
 import com.yoyomo.domain.application.application.mapper.AnswerMapper;
 import com.yoyomo.domain.application.domain.entity.Answer;
 import com.yoyomo.domain.application.domain.entity.Application;
-import com.yoyomo.domain.application.domain.model.ApplicationReply;
 import com.yoyomo.domain.application.domain.repository.mongo.AnswerRepository;
+import com.yoyomo.domain.application.domain.vo.ApplicationReply;
 import com.yoyomo.domain.application.exception.ApplicationReplySizeMismatchException;
 import com.yoyomo.domain.item.domain.entity.Item;
 import com.yoyomo.domain.item.domain.service.factory.ItemFactory;
